+++ R/tm_data_table.R
@@ -6,6 +6,8 @@
 #' @param variables_selected (\code{list}) a named list that says which variables should be
 #'   initially shown for particular dataset. Names in list should correspond with names provided in list `data()`.
 #'   If not specified for any dataset - first six variables from dataset will be shown.
+#' @param datasets_selected (\code{character}) a character vector that says which datasets should be
+#'   shown and in what order. Names in a vector have to correspond with datasets names.
 #'
 #' @export
 #'
@@ -22,12 +24,12 @@
 #'   ),
 #'   modules = root_modules(
 #'     tm_data_table(
-#'       variables_selected = list(ADSL  = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"))
+#'       variables_selected = list(ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"))
 #'     )
 #'   )
 #' )
 #' \dontrun{
-#'   shinyApp(app$ui, app$server)
+#' shinyApp(app$ui, app$server)
 #' }
 #'
 #' # two-datasets example
@@ -46,13 +48,18 @@
 #'   ),
 #'   modules = root_modules(
 #'     tm_data_table(
-#'       variables_selected = list(ADSL  = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
-#'                                 ADTTE = c("STUDYID", "USUBJID", "SUBJID", "SITEID",
-#'                                           "PARAM", "PARAMCD", "ARM", "ARMCD", "AVAL", "CNSR")))
+#'       variables_selected = list(
+#'         ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
+#'         ADTTE = c(
+#'           "STUDYID", "USUBJID", "SUBJID", "SITEID",
+#'           "PARAM", "PARAMCD", "ARM", "ARMCD", "AVAL", "CNSR"
+#'         )
+#'       )
+#'     )
 #'   )
 #' )
 #' \dontrun{
-#'   shinyApp(app$ui, app$server)
+#' shinyApp(app$ui, app$server)
 #' }
 #'
 #' # datasets: different subsets of long dataset
@@ -70,39 +77,100 @@
 #'   ),
 #'   modules = root_modules(
 #'     tm_data_table(
-#'       variables_selected = list(ADSL  = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
-#'                                 ADLB = c("STUDYID", "USUBJID", "SUBJID", "SITEID",
-#'                                          "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG")))
+#'       variables_selected = list(
+#'         ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
+#'         ADLB = c(
+#'           "STUDYID", "USUBJID", "SUBJID", "SITEID",
+#'           "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
+#'         )
+#'       )
+#'     )
 #'   )
 #' )
-#'
 #' \dontrun{
-#'   shinyApp(app$ui, app$server)
+#' shinyApp(app$ui, app$server)
+#' }
+#' # datasets: subsetting or changing order of datasets inside tm_data_table
+#' library(random.cdisc.data)
+#'
+#' ADSL <- radsl(cached = TRUE)
+#' ADLB <- radlb(cached = TRUE)
+#' ADTTE <- radtte(cached = TRUE)
+#'
+#' app <- init(
+#'   data = cdisc_data(
+#'     cdisc_dataset("ADSL", ADSL),
+#'     cdisc_dataset("ADLB", ADLB),
+#'     cdisc_dataset("ADTTE", ADTTE),
+#'     code = "ADSL <- radsl(cached = TRUE); ADLB <- radlb(cached = TRUE); ADTTE <- radtte(cached = TRUE);",
+#'     check = TRUE
+#'   ),
+#'   modules = root_modules(
+#'     tm_data_table(
+#'       variables_selected = list(
+#'         ADSL = c("STUDYID", "USUBJID", "SUBJID", "SITEID", "AGE", "SEX"),
+#'         ADLB = c(
+#'           "STUDYID", "USUBJID", "SUBJID", "SITEID",
+#'           "PARAM", "PARAMCD", "AVISIT", "AVISITN", "AVAL", "CHG"
+#'         )
+#'       ),
+#'       datasets_selected = c("ADLB", "ADSL")
+#'     )
+#'   )
+#' )
+#' \dontrun{
+#' shinyApp(app$ui, app$server)
 #' }
 #'
 tm_data_table <- function(label = "Data table",
-                          variables_selected = list()) {
+                          variables_selected = list(),
+                          datasets_selected = character(0)) {
   stopifnot(
     is_character_single(label),
     is.list(variables_selected),
-    `if`(length(variables_selected) > 0,
-         !is.null(names(variables_selected)), TRUE),
-    `if`(length(variables_selected) > 0,
-         all(vapply(names(variables_selected), is.character, FUN.VALUE = logical(1))), TRUE),
-    `if`(length(variables_selected) > 0,
-         all(vapply(names(variables_selected), nchar, FUN.VALUE = integer(1)) > 0), TRUE),
-    `if`(length(variables_selected) > 0,
-         all(vapply(variables_selected, is.character, FUN.VALUE = logical(1))), TRUE),
-    `if`(length(variables_selected) > 0,
-         all(vapply(variables_selected, length, FUN.VALUE = integer(1)) > 0), TRUE)
+    is_character_empty(datasets_selected) || is_character_vector(datasets_selected),
+
+    `if`(
+      length(variables_selected) > 0,
+      !is.null(names(variables_selected)), TRUE
+    ),
+    `if`(
+      length(variables_selected) > 0,
+      all(vapply(names(variables_selected), is.character, FUN.VALUE = logical(1))), TRUE
+    ),
+    `if`(
+      length(variables_selected) > 0,
+      all(vapply(names(variables_selected), nchar, FUN.VALUE = integer(1)) > 0), TRUE
+    ),
+    `if`(
+      length(variables_selected) > 0,
+      all(vapply(variables_selected, is.character, FUN.VALUE = logical(1))), TRUE
+    ),
+    `if`(
+      length(variables_selected) > 0,
+      all(vapply(variables_selected, length, FUN.VALUE = integer(1)) > 0), TRUE
+    ),
+
+    `if`(
+      length(datasets_selected) > 0,
+      all(vapply(datasets_selected, nchar, FUN.VALUE = integer(1)) > 0), TRUE
+    ),
+    `if`(
+      length(datasets_selected) > 0,
+      all(vapply(datasets_selected, is.character, FUN.VALUE = logical(1))), TRUE
+    )
   )
 
   module(
     label,
     server = srv_page_data_table,
     ui = ui_page_data_table,
-    filters = "all",
-    ui_args = list(selected = variables_selected)
+    filters = if_character_empty(datasets_selected, "all"),
+    server_args = list(datasets_selected = datasets_selected),
+    ui_args = list(
+      selected = variables_selected,
+      datasets_selected = datasets_selected
+    )
   )
 }
 
@@ -111,10 +179,12 @@ tm_data_table <- function(label = "Data table",
 #' @importFrom utils head
 ui_page_data_table <- function(id,
                                datasets,
-                               selected) {
+                               selected,
+                               datasets_selected) {
   ns <- NS(id)
 
-  datanames <- datasets$datanames()
+  datanames <- get_datanames_selected(datasets, datasets_selected)
+
   tagList(
     fluidRow(
       column(
@@ -187,13 +257,15 @@ ui_page_data_table <- function(id,
 srv_page_data_table <- function(input,
                                 output,
                                 session,
-                                datasets) {
-
+                                datasets,
+                                datasets_selected) {
   if_filtered <- reactive(as.logical(input$if_filtered))
   if_distinct <- reactive(as.logical(input$if_distinct))
 
+  datanames <- get_datanames_selected(datasets, datasets_selected)
+
   lapply(
-    datasets$datanames(),
+    datanames,
     function(x) {
       callModule(
         module = srv_data_table,
@@ -244,7 +316,6 @@ srv_data_table <- function(input,
                            dataname,
                            if_filtered,
                            if_distinct) {
-
   output$data_table <- DT::renderDataTable({
     variables <- input$variables
 
@@ -278,4 +349,21 @@ srv_data_table <- function(input,
       )
     )
   })
+}
+
+#' a tool for ui and server for getting datanames taking into account the datasets_selected vector
+#'
+#' @param datasets teal datasets object
+#' @param datasets_selected (\code{character}) a character vector that says which datasets should be
+#'   shown and in what order. Names in a vector have to correspond with datasets names.
+#' @return (\code{character}) a character vector
+get_datanames_selected <- function(datasets, datasets_selected) {
+  datanames <- datasets$datanames()
+
+  if (!is_character_empty(datasets_selected)) {
+    stopifnot(all(datasets_selected %in% datanames))
+    datanames <- datasets_selected
+  }
+
+  return(datanames)
 }